feat(subgraphs/flt): refactor
